search table view displaying expenses
