+++ budget/VC/Search/SearchExpensesViewController.swift
@@ -1,0 +1,181 @@
+//
+//  SearchEntriesViewController.swift
+//  budget
+//
+//  Created by Bradley Yin on 11/24/19.
+//  Copyright © 2019 bradleyyin. All rights reserved.
+//
+
+import UIKit
+import CoreData
+
+class SearchExpensesViewController: UIViewController {
+    var budgetController: BudgetController!
+    
+    var tableView: UITableView!
+    
+     lazy var fetchedResultsController: NSFetchedResultsController<Expense>? = {
+        let fetchRequest: NSFetchRequest<Expense> = Expense.fetchRequest()
+        fetchRequest.sortDescriptors = [NSSortDescriptor(key: "date", ascending: true)]
+        
+        let moc = CoreDataStack.shared.mainContext
+        let frc = NSFetchedResultsController(fetchRequest: fetchRequest, managedObjectContext: moc, sectionNameKeyPath: nil, cacheName: nil)
+        
+        frc.delegate = self
+        
+     do {
+         try frc.performFetch()
+     } catch {
+         fatalError("cant fetch category")
+     }
+        return frc
+    }()
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        configureTableView()
+        setConstraint()
+        // Do any additional setup after loading the view.
+    }
+    
+    private func configureTableView() {
+        let tableView = UITableView()
+        tableView.translatesAutoresizingMaskIntoConstraints = false
+        self.view.addSubview(tableView)
+        tableView.delegate = self
+        tableView.dataSource = self
+        tableView.backgroundColor = .clear
+        tableView.register(ExpenseTableViewCell.self, forCellReuseIdentifier: "ExpenseCell")
+        self.tableView = tableView
+    }
+    
+    private func setConstraint() {
+        tableView.topAnchor.constraint(equalTo: view.safeAreaLayoutGuide.topAnchor).isActive = true
+        tableView.bottomAnchor.constraint(equalTo: view.safeAreaLayoutGuide.bottomAnchor).isActive = true
+        tableView.leadingAnchor.constraint(equalTo: view.leadingAnchor).isActive = true
+        tableView.trailingAnchor.constraint(equalTo: view.trailingAnchor).isActive = true
+    }
+    
+    override func viewDidLayoutSubviews() {
+        super.viewDidLayoutSubviews()
+        setupUIColor()
+    }
+    
+    private func setupUIColor() {
+        if traitCollection.userInterfaceStyle == .light {
+            self.view.backgroundColor = .white
+        } else {
+            self.view.backgroundColor = .black
+        }
+    }
+    
+//    private func refreshFRC(keyword: String) {
+//        fetchedResultsController = nil
+//        fetchedResultsController = {
+//            let request: NSFetchRequest<Expense> = Expense.fetchRequest()
+//
+//            guard let date = date as NSDate? else { fatalError("cannot convert date for fetching") }
+//            let predicate = NSPredicate(format: "date == %@", date)
+//            let sortDescriptor = NSSortDescriptor(key: "name", ascending: true)
+//
+//
+//            request.predicate = predicate
+//            request.sortDescriptors = [sortDescriptor]
+//
+//            let moc = CoreDataStack.shared.mainContext
+//            let frc = NSFetchedResultsController(fetchRequest: request, managedObjectContext: moc, sectionNameKeyPath: nil, cacheName: nil)
+//
+//            frc.delegate = self
+//
+//         do {
+//             try frc.performFetch()
+//         } catch {
+//             fatalError("cant fetch expense")
+//         }
+//            return frc
+//        }()
+//
+//    }
+
+}
+
+extension SearchExpensesViewController: UITableViewDelegate, UITableViewDataSource {
+    func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
+        return fetchedResultsController?.fetchedObjects?.count ?? 0
+    }
+    
+    func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
+        guard let cell = tableView.dequeueReusableCell(withIdentifier: "ExpenseCell", for: indexPath) as? ExpenseTableViewCell else {
+            fatalError("cant make ExpenseTableViewCell")
+        }
+        cell.expense = fetchedResultsController?.object(at: indexPath)
+        
+        return cell
+    }
+    
+    func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
+        
+        
+        if indexPath.section == 0 {
+            let singleDayDetailVC = SingleExpenseDetailViewController()
+            singleDayDetailVC.expense = fetchedResultsController?.object(at: indexPath)
+            singleDayDetailVC.budgetController = budgetController
+            tableView.deselectRow(at: indexPath, animated: false)
+            navigationController?.pushViewController(singleDayDetailVC, animated: true)
+        }
+        
+    }
+    
+    
+}
+
+extension SearchExpensesViewController: NSFetchedResultsControllerDelegate {
+    func controllerWillChangeContent(_ controller: NSFetchedResultsController<NSFetchRequestResult>) {
+        tableView.beginUpdates()
+    }
+    
+    func controllerDidChangeContent(_ controller: NSFetchedResultsController<NSFetchRequestResult>) {
+        tableView.endUpdates()
+    }
+    
+    func controller(_ controller: NSFetchedResultsController<NSFetchRequestResult>,
+                    didChange anObject: Any,
+                    at indexPath: IndexPath?,
+                    for type: NSFetchedResultsChangeType,
+                    newIndexPath: IndexPath?) {
+        switch type {
+        case .insert:
+            guard let newIndexPath = newIndexPath else { return }
+            tableView.insertRows(at: [newIndexPath], with: .automatic)
+        case .delete:
+            guard let indexPath = indexPath else { return }
+            tableView.deleteRows(at: [indexPath], with: .automatic)
+        case .move:
+            guard let indexPath = indexPath, let newIndexPath = newIndexPath else { return }
+            tableView.moveRow(at: indexPath, to: newIndexPath)
+        case .update:
+            guard let indexPath = indexPath else { return }
+            tableView.reloadRows(at: [indexPath], with: .automatic)
+        @unknown default:
+            fatalError("new cases for fetch result controller type")
+        }
+    }
+    
+    func controller(_ controller: NSFetchedResultsController<NSFetchRequestResult>,
+                    didChange sectionInfo: NSFetchedResultsSectionInfo,
+                    atSectionIndex sectionIndex: Int,
+                    for type: NSFetchedResultsChangeType) {
+        
+        let sectionsIndexSet = IndexSet(integer: sectionIndex)
+        
+        
+        switch type {
+        case .insert:
+            tableView.insertSections(sectionsIndexSet, with: .automatic)
+        case .delete:
+            tableView.deleteSections(sectionsIndexSet, with: .automatic)
+        default:
+            break
+        }
+    }
+}
